requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
google-ads>=22.1.0
Pillow>=10.0.0
ShopifyAPI>=12.7.0
python-unsplash>=1.1.0
aiohttp>=3.9.1 
streamlit>=1.25.0
pandas>=2.0.2